requirements.txt
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
cycler==0.10.0
dash==1.7.0
dash-core-components==1.6.0
dash-html-components==1.0.2
dash-renderer==1.2.2
dash-table==4.5.1
decorator==4.4.1
env==0.1.0
Flask==1.1.1
Flask-Compress==1.4.0
Flask-SQLAlchemy==2.4.1
future==0.18.2
gunicorn==20.0.4
idna==2.8
imageio==2.6.1
imutils==0.5.3
ipython==7.11.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
networkx==2.4
numpy==1.18.0
pandas==0.25.3
parso==0.5.2
pickleshare==0.7.5
Pillow==6.2.1
plotly==4.4.1
progressbar==2.5
prompt-toolkit==3.0.2
psycopg2==2.8.4
Pygments==2.5.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
requests==2.22.0
retrying==1.3.3
scikit-image==0.16.2
scipy==1.4.1
six==1.13.0
SQLAlchemy==1.3.12
tqdm==4.41.1
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.8
Werkzeug==0.16.0
